appimagetool: calculate size via _runtime_binary_start and _runtime_binary_end.
On my Ubuntu yakkety x86_64 sytem, the _runtime_binary_size
symbol is unreliable. For non-ASLR, it points somewhere after
0x555555554000 (where appimagetool is loaded). And with ASLR,
it's... random.

So, instead of relying on the address of _runtime_binary_size
being the size of the runtime blob, calculate the actual size
by using _runtime_binary_start and _runtime_binary_end.