getsection: make use of Elf32_Ehdr and Elf32_Shdr on 32-bit systems
Using the 64-bit counter parts causes the location of the embedded ELF
section for `--updateinformation` to be calculated incorrectly, and
eventually causing a segmentation fault.

Notice this commit relies on GCC specific defines (`__i386__`), and
defaults to 64-bit if 32-bit was not detected.

See: https://github.com/probonopd/AppImageKit/pull/307
Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>