Implement platform check depending on pointer size
Implements #400.

Instead of adding all supported CPU architectures by hand, the check is
now performed against the pointer size. This should be sufficient for
the purpose. The preprocessor will issue an error for unknown platforms.

Tested with gcc and clang. The check doesn't have to be super portable,
according to my research the new check works with all relevant Linux
systems and compilers. It is probably "portable enough (TM)".