Introduce new --sign-key argument
A separate argument makes it easier to be able to specify a key for
signing the key and exporting the key afterwards. As --sign-args may
contain additional arguments which would not work when using
gpg[2] --export, introducing a separate key argument that is then
passed to both these gpg[2] calls makes more sense.