Make sure to export only a single key and embed it
The normal behavior of gpg[2] is to export all keys unless an ID is
explicitly passed. Therefore, to allow appimagetool to work on systems
with multiple pubkeys in the users' keyrings, the key ID is read from
the signature using gpg[2] --list-packets <.asc file>, and only that
key is exported.

This is a fix when no --sign-key is passed explicitly, causing gpg[2]
to sign with the default key.