Add appimage_registered_desktop_file_path
This new function uses glob() to find an AppImage's desktop file after
registration in XDG_DATA_HOME. As the MD5 is supposed to be unique
(collision resistant for our use case), this should work fine.
Instead of searching the AppImage's contents for the missing filename
component, it is sufficient to just build a pattern for glob() using
the MD5 hash of the AppImage, using a wildcard character in place of
the missing component. The glob() call should then return a single
file path, which is returned by this function.