Remove appimaged specific desktop integration code from libappimage
libappimage users do not want to use appimaged specific desktop actions
etc. This PR guards these code snippets with a compiler definition
that is only activated for appimaged builds.

This PR should be considered a quick fix for the issue. In the long
term, the code needs to be moved to appimaged's source files.
Desktop actions can easily be added to the desktop file since the
introduction of appimage_registered_desktop_file_path, so this
shouldn't be much work.