Fix nasty bug in memory management when VERSION is set
The destination pointer was set to a pointer of a variable on the stack,
which, as soon as the variable got out of scope, became invalid.
However, the value existed under this address until the next char array
of size PATH_MAX was allocated on stack. The value the destination
pointer pointed to has then been the appdata file's path. Therefore,
an AppImage was created under that path.

The fix was simple: just create a new memory buffer by using strdup().
As the memory is allocated on the heap, the pointer remains valid
until the memory is free'd explicitly.

It appears to me like nobody has used the VERSION environment variable
before to have appimagetool choose the right filename. Even
linuxdeployqt chooses the filename based on VERSION and passes it
explicitly to appimagetool. Therefore, it hasn't been discovered since
@probonopd introduced the VERSION stuff in 2016.