Avoid processing non-localized strings more than once
NULL values refer to the non-localized version of the keys. By default,
the array storing the locales contains only NULL values, and
the counter is set to 1. This makes sure that the non-localized string
is processed. Now, if the loop runs over the entire array, unless
there's <array size> localized entries, the non-localized string will
be processed more than once. This commit fixes this behavior.