Avoid injecting the current directory into paths
If LD_LIBRARY_PATH was previously not set, it could end up as:

    LD_LIBRARY_PATH=.../lib64/:

which might result in querying the current working directory for
libraries. Use the "Use Alternative Value" parameter expansion (POSIX
Shell standard) to expand to an empty value (if the value is empty or
unset) and to `:$LD_LIBRARY_PATH` otherwise.

This patch was created with this eye-bleeding command to avoid typos:

    sed -r 's/:"\$\{([A-Z_]+)\}"$/"${\1:+:$\1}"/' resources/AppRun