Only mount tmpfs to /tmp if it's not mounted yet
This is a logical error that renders the script non-working on Debian for instance. If grep returns true (i.e. /tmp is already mounted as tmpfs) lazy evaluation should skip the right part. If it returns false then we should mount /tmp. With && the logic is exactly the opposite. How was it working before?