Allow users to require signing
Currently, if signing fails, appimagetool just reports the error but continues to build the AppImage. This is not ideal for automated environments which want to make sure the resulting AppImages have been signed correctly.

This commit adds support for the APPIMAGETOOL_FORCE_SIGN environment variable, which makes appimagetool abort in case signing fails.