Log when APPIMAGETOOL_SIGN_PASSPHRASE is set
There appear to be cases in which a passphrase is set (and appimagetool
correctly changes the pinentry mode to loopback) in which the gpg-agent
never asks for a password. In such a scenario, the logging within our
passphrase callback is never used, making it impossible to tell where
the problem is exactly. With this change, the log will show whether or
not the passphrase env var was found.